backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.0
slowapi==0.1.9
python-socketio==5.10.0
python-engineio==4.7.1
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4
Pillow==10.1.0